Fix issue #1267: Update the typography of the Gallery title (#1270)
* Fix issue #1267: Update the typography of the Gallery title

* Fix title size after review

Anne already solved the task by aligning with the desktop and mobile version of the design, but we talked and agreed that in order for tablet sized screens to look good too we would hold on the Gallery title size change until user is on mobile screen size. 

Good job!